Finished untested hyper heuristic implementation.

+++ src/com/ai/problems/min_set/records/TimeObjectiveValue.java
@@ -4,5 +4,5 @@ package com.ai.problems.min_set.records;
  * @param Time Time it took to run the heuristic, relative.
  * @param ObjectiveValue An objective value for the heuristic. Can be a change from the previous, or an absolute measure.
  */
-public record TimeObjectiveValue(double Time, int ObjectiveValue) {
+public record TimeObjectiveValue(long Time, int ObjectiveValue) {
 }
